restructured project, no change to existing features
