+++ src/routes/posts.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
-const verify = require('./verifyToken.js');
-const User = require('../models/User.js');
+const verify = require('../utils/verifyToken');
+const User = require('../models/User');
 
 // a private route that check for jwt
 router.get('/', verify, async (req, res) => {
