+++ src/utils/hashHandler.js
