+++ src/utils/tokenGenerator.js
